requirements.txt
transformers
torch
Pillow
scipy
timm
